+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 Notable changes to this project.
 
+## [dev]
+- add `check_round_open` to check if there is a currently an active round
+
 ## [2.12.5] - 2022-10-24
 - add fncV3 to `daily_submissions_performances`
 - add `TC` to `SignalsAPI.get_leaderboard`